requirements.txt
numpy~=1.24.2
sounddevice~=0.4.5
matplotlib~=3.7.0
torch~=1.13.1
transformers~=4.26.1
scipy~=1.10.0
PySide6~=6.4.2
ffmpeg-python~=0.2.0